+++ src/core/user-access/scheduled-deck/scheduled-deck.repository.ts
@@ -1,0 +1,7 @@
+import { ScheduledDeck } from './scheduled-deck.entity';
+
+export interface ScheduledDeckRepository {
+  insert(scheduledDeck: ScheduledDeck): Promise<void>;
+
+  findAllByUser(userId: string): Promise<ScheduledDeck[]>;
+}